requirements.txt
Django>=3.0,<4.0
djangorestframework==3.14.0
psycopg2>=2.8
pyyaml
uritemplate
drf-spectacular